// tutorials/back/api/persistence/types.ts
interface Pizza {
  id: number;
  title: string;
  content: string;
}

interface PizzaToUpdate {
  title?: string;
  content?: string;
}

interface Drink {
  id: number;
  title: string;
  image: string;
  volume: number;
  price: number;
}

type NewPizza = Omit<Pizza, "id">;
type NewDrink = Omit<Drink, "id">;







export type { Pizza, NewPizza, PizzaToUpdate, Drink, NewDrink};